[clang-tidy] Improve the matcher to convert method calls to function calls
I think this is strictly an improvement on the last commit, but I'm not
sure, so I leave it as a separate commit.

This should handle all cases in C++ code outside of macros. There is a
variable where one can add or remove an `llvm::` prefix to all
conversions. This change also replaces dyn_cast_or_null with
dyn_cast_if_present and handles PointerUnion methods which sometimes
call other underlying functions.