[mlir] Mark cast like methods as deprecated where functional variant exists
Classes that have an associate ClassInfo class can use the functional
variant and have had all call sites migrated at this point.

dyn_cast is not deprecated because TypeSwitch has a specialized method
for classes that define a dyn_cast method, and thus result in a callsite
for the deprecated method. This cannot simply be removed because there
still exist classes that define the methods but not a ClassInfo for
casting.